Big ECS refactoring. Fix entity relationship serialization and deserialization
