+++ Cardia/src/Cardia/ECS/Scene.cpp
@@ -65,9 +65,17 @@ namespace Cardia
 		return entity;
 	}
 
-	void Scene::DestroyEntity(entt::entity entity)
+	void Scene::DestroyEntity(const Entity entity)
 	{
-		m_Registry.destroy(entity);
+		const auto& relationship = entity.GetComponent<Component::Relationship>();
+		auto current = Entity(relationship.FirstChild, this);
+		while (current.IsValid())
+		{
+			const auto next = current.GetComponent<Component::Relationship>().NextSibling;
+			DestroyEntity(current);
+			current = Entity(next, this);
+		}
+		m_Registry.destroy(entity.Handle());
 	}
 
 	void Scene::OnRuntimeRender(VkCommandBuffer commandBuffer)
@@ -157,16 +165,20 @@ namespace Cardia
 
 	Entity Scene::GetEntityByUUID(const UUID& uuid)
 	{
-		auto view = m_Registry.view<Component::ID>();
-		Entity result;
-		for (auto entity : view)
+		const auto view = m_Registry.view<Component::ID>();
+		for (const auto entity : view)
 		{
 			auto& idComponent = view.get<Component::ID>(entity);
 			if (idComponent.Uuid == uuid) {
-				result = Entity(entity, this);
+				return {entity, this};
 			}
 		}
-		return result;
+		return {entt::null, this};
+	}
+
+	bool Scene::IsEntityValid(entt::entity entity) const
+	{
+		return m_Registry.valid(entity);
 	}
 
 	void Scene::OnRuntimeStop()
@@ -217,6 +229,22 @@ namespace Cardia
 		{
 			Json::Value currentEntityNode(Json::objectValue);
 			SerializeAllComponents(SerializableComponents{}, m_Registry, currentEntityNode, entity);
+
+			// Serialize Relationships here as they need scene context to be deserialized
+			const auto& relationship = m_Registry.get<Component::Relationship>(entity);
+			Json::Value toMerge(Json::objectValue);
+			auto& relationshipNode = toMerge["Relationship"];
+			relationshipNode["ChildCount"] = relationship.ChildCount;
+			if (m_Registry.valid(relationship.Parent))
+				relationshipNode["Parent"] = m_Registry.get<Component::ID>(relationship.Parent).Uuid.ToString();
+			if (m_Registry.valid(relationship.FirstChild))
+				relationshipNode["FirstChild"] = m_Registry.get<Component::ID>(relationship.FirstChild).Uuid.ToString();
+			if (m_Registry.valid(relationship.PreviousSibling))
+				relationshipNode["PreviousSibling"] = m_Registry.get<Component::ID>(relationship.PreviousSibling).Uuid.ToString();
+			if (m_Registry.valid(relationship.NextSibling))
+				relationshipNode["NextSibling"] = m_Registry.get<Component::ID>(relationship.NextSibling).Uuid.ToString();
+
+			MergeJson(currentEntityNode, toMerge);
 			entitiesNode.append(currentEntityNode);
 		}
 
@@ -247,10 +275,37 @@ namespace Cardia
 
 		Scene scene(AssetsManager::Instance().GetRenderer(), std::string("Deserialized Scene"));
 
+		// Deserialize all serializable components
 		for (auto& entityNode : root["Entities"])
 		{
 			const auto entity = scene.m_Registry.create();
 			DeserializeAndAssignAllComponents(SerializableComponents{}, entityNode, scene.m_Registry, entity);
+		}
+
+		// Second pass to deserialize things that need scene context
+		for (auto& entityNode : root["Entities"])
+		{
+			std::optional<Component::ID> id;
+			if (!((id = Component::ID::Deserialize(entityNode))))
+				continue;
+
+			auto entity = scene.GetEntityByUUID(id->Uuid);
+
+			auto& relationship = entity.AddComponent<Component::Relationship>();
+
+			if (!entityNode.isMember("Relationship"))
+				continue;
+			
+			auto& relationshipNode = entityNode["Relationship"];
+			relationship.ChildCount = relationshipNode["ChildCount"].asInt();
+			if (relationshipNode.isMember("Parent"))
+				relationship.Parent = scene.GetEntityByUUID(UUID::FromString(relationshipNode["Parent"].asString())).Handle();
+			if (relationshipNode.isMember("FirstChild"))
+				relationship.FirstChild = scene.GetEntityByUUID(UUID::FromString(relationshipNode["FirstChild"].asString())).Handle();
+			if (relationshipNode.isMember("PreviousSibling"))
+				relationship.PreviousSibling = scene.GetEntityByUUID(UUID::FromString(relationshipNode["PreviousSibling"].asString())).Handle();
+			if (relationshipNode.isMember("NextSibling"))
+				relationship.NextSibling = scene.GetEntityByUUID(UUID::FromString(relationshipNode["NextSibling"].asString())).Handle();
 		}
 
 		return scene;
@@ -271,17 +326,17 @@ namespace Cardia
 			auto& parentRelationship = m_Registry.get<Component::Relationship>(parent);
 			auto current = parentRelationship.FirstChild;
 
-			if (current == entt::null)
+			if (!m_Registry.valid(current))
 				parentRelationship.FirstChild = entity.Handle();
 			else {
-				while (m_Registry.get<Component::Relationship>(current).NextSibling != entt::null)
+				while (m_Registry.valid(m_Registry.get<Component::Relationship>(current).NextSibling))
 				{
 					current = m_Registry.get<Component::Relationship>(current).NextSibling;
 				}
 				m_Registry.get<Component::Relationship>(current).NextSibling = entity.Handle();
 				relationship.PreviousSibling = current;
-				parentRelationship.ChildCount++;
 			}
+			parentRelationship.ChildCount++;
 		}
 	}
